Added emails

+++ src/api/api_v1/endpoints/confirm_email.py
@@ -10,9 +10,9 @@ from schemas.authorization import AuthorizationResponse
 from models.confirm_email_token import ConfirmEmailToken
 from models.pre_user import PreUser
 from models.user import User
-from core.security import AuthLoginForm, verify_auth_form, create_refresh_token, verify_refresh_token, get_owner_refresh_token, create_access_token
+from core.security import create_email_confirmation_token
 from core.config import settings
-
+from core.email import send_confirmation_email
 router = APIRouter()
 
 """[Confirmar email]
@@ -33,11 +33,13 @@ async def confirm_email(token : str):
     print(token_orm.valid_until)
     print(datetime.utcnow())
     if token_orm is None or token_orm.valid_until < datetime.utcnow():
+        if not token_orm:
+            await token_orm.delete()
         raise HTTPException(status_code=400, detail={'error':'invalid_grant','error_description':'Invalid token'})
     
     pre_user : PreUser = await PreUser.get_or_404(token_orm.pre_user_id)
 
-    await token_orm.delete()
+    await ConfirmEmailToken.delete.where(ConfirmEmailToken.pre_user_id == pre_user.id).gino.status()
     await User.create(
             fullname = pre_user.fullname,
             email = pre_user.email,
@@ -49,3 +51,11 @@ async def confirm_email(token : str):
     await pre_user.delete()
 
     return {'success':'email_confirmed', 'success_description':'Email confirmed with success'}
+
+@router.post('/') 
+async def request_confirm_email(pre_user_id : int):
+    pre_user : PreUser = await PreUser.get_or_404(pre_user_id)
+    token :str = await create_email_confirmation_token(pre_user)
+    
+    await send_confirmation_email(pre_user.email, token)
+    return token
